# lib/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Robux Donation Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RobloxUser {
  id: number;
  name: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface Gamepass {
  id: number;
  name: string;
  price: number;
  /** @nullable */
  iconUrl?: string | null;
  /** @nullable */
  sellerName?: string | null;
  gameId: number;
  /** @nullable */
  gameName?: string | null;
}

export interface GamepassList {
  gamepasses: Gamepass[];
  cached: boolean;
  fetchedAt: string;
}