# lib/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Robux Donation Hub API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: roblox
    description: Roblox proxy endpoints
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /user/{username}:
    get:
      operationId: getRobloxUser
      tags: [roblox]
      summary: Look up a Roblox user by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RobloxUser"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /gamepasses/{userId}:
    get:
      operationId: getGamepasses
      tags: [roblox]
      summary: Fetch all on-sale gamepasses for a Roblox user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of gamepasses sorted by price ascending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GamepassList"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    RobloxUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
      required:
        - id
        - name
        - displayName

    Gamepass:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: integer
        iconUrl:
          type: ["string", "null"]
        sellerName:
          type: ["string", "null"]
        gameId:
          type: integer
        gameName:
          type: ["string", "null"]
      required:
        - id
        - name
        - price
        - gameId

    GamepassList:
      type: object
      properties:
        gamepasses:
          type: array
          items:
            $ref: "#/components/schemas/Gamepass"
        cached:
          type: boolean
        fetchedAt:
          type: string
      required:
        - gamepasses
        - cached
        - fetchedAt