# lib/lib/api-zod/src/generated/types/gamepass.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Robux Donation Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Gamepass {
  id: number;
  name: string;
  price: number;
  /** @nullable */
  iconUrl?: string | null;
  /** @nullable */
  sellerName?: string | null;
  gameId: number;
  /** @nullable */
  gameName?: string | null;
}