# lib/lib/api-zod/src/generated/types/gamepassList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Robux Donation Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { Gamepass } from './gamepass';

export interface GamepassList {
  gamepasses: Gamepass[];
  cached: boolean;
  fetchedAt: string;
}