# lib/lib/api-zod/src/generated/types/robloxUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Robux Donation Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface RobloxUser {
  id: number;
  name: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  description?: string | null;
}